Prepare maple_cli seperation
