+++ crates/maple_cli/src/command/grep/mod.rs
@@ -2,29 +2,16 @@ mod forerunner;
 mod live_grep;
 
 use crate::app::Params;
-use crate::cache::Digest;
-use crate::process::ShellCommand;
+use crate::tools::ripgrep::{refresh_cache, rg_shell_command};
 use anyhow::Result;
 use clap::Parser;
 use filter::{ParallelSource, SequentialSource};
 use matcher::MatchScope;
-use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::path::PathBuf;
 use subprocess::Exec;
 
 pub use self::forerunner::RipGrepForerunner;
 pub use self::live_grep::LiveGrep;
-
-const RG_ARGS: &[&str] = &[
-    "rg",
-    "--column",
-    "--line-number",
-    "--no-heading",
-    "--color=never",
-    "--smart-case",
-    "",
-    ".",
-];
 
 // Ref https://github.com/liuchengxu/vim-clap/issues/533
 // Now `.` is pushed to the end for all platforms due to https://github.com/liuchengxu/vim-clap/issues/711.
@@ -141,64 +128,11 @@ impl Grep {
     }
 }
 
-// Used for creating the cache in async context.
-#[derive(Debug, Clone, Hash)]
-pub struct RgTokioCommand {
-    shell_cmd: ShellCommand,
-}
-
-impl RgTokioCommand {
-    pub fn new(dir: PathBuf) -> Self {
-        let shell_cmd = ShellCommand::new(RG_EXEC_CMD.into(), dir);
-        Self { shell_cmd }
-    }
-
-    pub fn cache_digest(&self) -> Option<Digest> {
-        self.shell_cmd.cache_digest()
-    }
-
-    pub async fn create_cache(self) -> Result<Digest> {
-        let cache_file = self.shell_cmd.cache_file_path()?;
-
-        let std_cmd = rg_command(&self.shell_cmd.cwd);
-        let mut tokio_cmd = tokio::process::Command::from(std_cmd);
-        crate::process::tokio::write_stdout_to_file(&mut tokio_cmd, &cache_file).await?;
-
-        let digest = crate::cache::store_cache_digest(self.shell_cmd.clone(), cache_file)?;
-
-        Ok(digest)
-    }
-}
-
-pub fn rg_command<P: AsRef<Path>>(dir: P) -> Command {
-    // Can not use StdCommand as it joins the args which does not work somehow.
-    let mut cmd = Command::new(RG_ARGS[0]);
-    // Do not use --vimgrep here.
-    cmd.args(&RG_ARGS[1..]).current_dir(dir);
-    cmd
-}
-
-#[inline]
-pub fn rg_shell_command<P: AsRef<Path>>(dir: P) -> ShellCommand {
-    ShellCommand::new(RG_EXEC_CMD.into(), PathBuf::from(dir.as_ref()))
-}
-
-pub fn refresh_cache(dir: impl AsRef<Path>) -> Result<Digest> {
-    let shell_cmd = rg_shell_command(dir.as_ref());
-    let cache_file_path = shell_cmd.cache_file_path()?;
-
-    let mut cmd = rg_command(dir.as_ref());
-    crate::process::write_stdout_to_file(&mut cmd, &cache_file_path)?;
-
-    let digest = crate::cache::store_cache_digest(shell_cmd, cache_file_path)?;
-
-    Ok(digest)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::process::tokio::TokioCommand;
+    use crate::tools::ripgrep::RgTokioCommand;
     use itertools::Itertools;
     use std::time::Instant;
