+++ crates/maple_cli/src/tools/ripgrep/mod.rs
@@ -2,6 +2,8 @@ mod default_types;
 mod jsont;
 mod stats;
 
+use crate::cache::Digest;
+use crate::process::ShellCommand;
 use crate::utils::display_width;
 use anyhow::Result;
 use once_cell::sync::Lazy;
@@ -9,6 +11,8 @@ use std::borrow::Cow;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::ops::Range;
+use std::path::{Path, PathBuf};
+use std::process::Command;
 
 pub use self::jsont::{Match, Message, SubMatch};
 
@@ -250,4 +254,74 @@ impl Match {
         let indices = self.match_indices_for_dumb_jump(offset, word);
         (formatted, indices)
     }
+}
+
+const RG_ARGS: &[&str] = &[
+    "rg",
+    "--column",
+    "--line-number",
+    "--no-heading",
+    "--color=never",
+    "--smart-case",
+    "",
+    ".",
+];
+
+// Ref https://github.com/liuchengxu/vim-clap/issues/533
+// Now `.` is pushed to the end for all platforms due to https://github.com/liuchengxu/vim-clap/issues/711.
+pub const RG_EXEC_CMD: &str =
+    "rg --column --line-number --no-heading --color=never --smart-case '' .";
+
+// Used for creating the cache in async context.
+#[derive(Debug, Clone, Hash)]
+pub struct RgTokioCommand {
+    shell_cmd: ShellCommand,
+}
+
+impl RgTokioCommand {
+    pub fn new(dir: PathBuf) -> Self {
+        let shell_cmd = ShellCommand::new(RG_EXEC_CMD.into(), dir);
+        Self { shell_cmd }
+    }
+
+    pub fn cache_digest(&self) -> Option<Digest> {
+        self.shell_cmd.cache_digest()
+    }
+
+    pub async fn create_cache(self) -> Result<Digest> {
+        let cache_file = self.shell_cmd.cache_file_path()?;
+
+        let std_cmd = rg_command(&self.shell_cmd.cwd);
+        let mut tokio_cmd = tokio::process::Command::from(std_cmd);
+        crate::process::tokio::write_stdout_to_file(&mut tokio_cmd, &cache_file).await?;
+
+        let digest = crate::cache::store_cache_digest(self.shell_cmd.clone(), cache_file)?;
+
+        Ok(digest)
+    }
+}
+
+pub fn rg_command<P: AsRef<Path>>(dir: P) -> Command {
+    // Can not use StdCommand as it joins the args which does not work somehow.
+    let mut cmd = Command::new(RG_ARGS[0]);
+    // Do not use --vimgrep here.
+    cmd.args(&RG_ARGS[1..]).current_dir(dir);
+    cmd
+}
+
+pub fn refresh_cache(dir: impl AsRef<Path>) -> Result<Digest> {
+    let shell_cmd = rg_shell_command(dir.as_ref());
+    let cache_file_path = shell_cmd.cache_file_path()?;
+
+    let mut cmd = rg_command(dir.as_ref());
+    crate::process::write_stdout_to_file(&mut cmd, &cache_file_path)?;
+
+    let digest = crate::cache::store_cache_digest(shell_cmd, cache_file_path)?;
+
+    Ok(digest)
+}
+
+#[inline]
+pub fn rg_shell_command<P: AsRef<Path>>(dir: P) -> ShellCommand {
+    ShellCommand::new(RG_EXEC_CMD.into(), PathBuf::from(dir.as_ref()))
 }
